+++ release/scripts/startup/abler/general.py
@@ -105,53 +105,6 @@ class ImportOperator(bpy.types.Operator, ImportHelper):
         return {"FINISHED"}
 
 
-class ImportFBXOperator(bpy.types.Operator, ImportHelper):
-    """Import FBX file according to the current settings"""
-
-    bl_idname = "acon3d.import_fbx"
-    bl_label = "Import FBX"
-    bl_translation_context = "*"
-
-    filter_glob: bpy.props.StringProperty(default="*.fbx", options={"HIDDEN"})
-
-    def execute(self, context):
-
-        for obj in bpy.data.objects:
-            obj.select_set(False)
-
-        FILEPATH = self.filepath
-
-        filename = os.path.basename(FILEPATH)
-        col_imported = bpy.data.collections.new("[FBX] " + filename.replace(".fbx", ""))
-
-        col_layers = bpy.data.collections.get("Layers")
-        if not col_layers:
-            col_layers = bpy.data.collections.new("Layers")
-            context.scene.collection.children.link(col_layers)
-
-        bpy.ops.import_scene.fbx(filepath=FILEPATH)
-        for obj in bpy.context.selected_objects:
-            if obj.name in bpy.context.scene.collection.objects:
-                bpy.context.scene.collection.objects.unlink(obj)
-            for c in bpy.data.collections:
-                if obj.name in c.objects:
-                    c.objects.unlink(obj)
-            col_imported.objects.link(obj)
-
-        # put col_imported in l_exclude
-        col_layers.children.link(col_imported)
-        added_l_exclude = context.scene.l_exclude.add()
-        added_l_exclude.name = col_imported.name
-        added_l_exclude.value = True
-
-        # create group
-        bpy.ops.acon3d.create_group()
-        # apply AconToonStyle
-        materials_setup.applyAconToonStyle()
-
-        return {"FINISHED"}
-
-
 class ToggleToolbarOperator(bpy.types.Operator):
     """Toggle toolbar visibility"""
 
@@ -269,10 +222,6 @@ class Acon3dImportPanel(bpy.types.Panel):
         row.operator("acon3d.save", text="Save")
         row.operator("acon3d.save_as", text="Save As...")
 
-        layout.separator()
-        row = layout.row()
-        row.operator("acon3d.import_fbx", text="Import FBX")
-
         row = layout.row()
         prefs = context.preferences
         view = prefs.view
@@ -305,7 +254,6 @@ classes = (
     ApplyToonStyleOperator,
     FileOpenOperator,
     FlyOperator,
-    ImportFBXOperator,
     SaveOperator,
     SaveAsOperator,
 )
